finished details page basic styling
