+++ pages/_app.tsx
@@ -7,7 +7,7 @@ import { createTheme, ThemeProvider } from '@mui/material';
 import CssBaseline from '@mui/material/CssBaseline';
 import type { AppProps } from 'next/app';
 import * as React from 'react';
-import Navbar from '../components/NavBar';
+import Screen from '../components/Screen';
 import '../styles/globals.css';
 import theme from '../styles/theme/Theme';
 import createEmotionCache from '../utility/createEmotionCache';
@@ -27,11 +27,12 @@ const MyApp: React.FunctionComponent<MyAppProps> = (props) => {
   return (
     <CacheProvider value={emotionCache}>
       <ThemeProvider theme={defaultTheme}>
-        <CssBaseline />
-        <Navbar />
-        <div>
-          <Component {...pageProps} />
-        </div>
+        <AuthProvider>
+          <CssBaseline />
+          <Screen>
+            <Component {...pageProps} />
+          </Screen>
+        </AuthProvider>
       </ThemeProvider>
     </CacheProvider>
   );
