+++ pages/details.tsx
@@ -2,19 +2,9 @@ import { User } from '../utility/types/user';
 import React, { useState } from 'react';
 import { useAuth } from '../utility/hooks/authentication';
 import { SelectListHook } from '../utility/hooks/selectList';
-import {
-  Button,
-  FormControl,
-  Grid,
-  Paper,
-  TextField,
-  ThemeProvider,
-  Typography,
-} from '@mui/material';
+import { Button, Grid, Paper, TextField, Typography } from '@mui/material';
 import { SelectChipList } from '../components/SelectChipList';
-import { ArrowRight } from '@mui/icons-material';
 import theme from '../styles/theme/Theme';
-import { flexbox } from '@mui/system';
 
 export default function Details() {
   const { addDetails, user } = useAuth();
@@ -75,127 +65,122 @@ export default function Details() {
     });
   }
 
-  /* TODO: Add go back button */
-
   return (
-    <ThemeProvider theme={theme}>
-      <div
+    <div style={$wrapper}>
+      <Paper
+        elevation={3}
         style={{
-          width: '100%',
-          height: '100%',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          padding: 10,
+          width: '90%',
+          maxWidth: '600px',
+          background: theme.palette.background.paper,
+          padding: 20,
+          paddingBottom: 40,
         }}
       >
-        <Paper
-          elevation={3}
-          style={{
-            width: '90%',
-            maxWidth: '500px',
-            background: theme.palette.primary.main,
-            padding: 20,
-            paddingBottom: 40,
-          }}
+        <Grid
+          container
+          direction="column"
+          justifyContent="center"
+          alignItems="center"
         >
-          <Grid
-            container
-            direction="column"
-            justifyContent="center"
-            alignItems="center"
-          >
-            <div style={$container}>
-              <Typography
-                variant="h4"
-                style={{
-                  textAlign: 'center',
-                  fontWeight: 500,
-                  color: theme.palette.secondary.main,
-                  padding: 5,
-                }}
-              >
-                details
-              </Typography>
-              <Typography
-                style={{
-                  textAlign: 'center',
-                  color: theme.palette.secondary.main,
-                  paddingBottom: 15,
-                }}
-              >
-                please enter some additional details.
-              </Typography>
+          <div style={$container}>
+            <Typography
+              variant="h4"
+              style={{
+                textAlign: 'center',
+                fontWeight: 500,
+                color: theme.palette.secondary.main,
+                padding: 5,
+              }}
+            >
+              details
+            </Typography>
+            <Typography
+              style={{
+                textAlign: 'center',
+                color: theme.palette.secondary.main,
+                paddingBottom: 15,
+              }}
+            >
+              please enter some additional details.
+            </Typography>
 
-              <TextField
-                color="secondary"
-                sx={{ width: '100%', marginTop: 2, marginBottom: 2 }}
-                error={isNameInvalid}
-                placeholder={isNameInvalid ? 'missing name' : undefined}
-                value={details.name}
-                label="name"
-                onChange={(e) =>
-                  sDetails((details) => ({
-                    ...details,
-                    name: e.target.value,
-                  }))
-                }
-              />
+            <TextField
+              color="secondary"
+              sx={{ width: '100%', marginTop: 2, marginBottom: 2 }}
+              error={isNameInvalid}
+              placeholder={isNameInvalid ? 'missing name' : undefined}
+              value={details.name}
+              label="name"
+              onChange={(e) =>
+                sDetails((details) => ({
+                  ...details,
+                  name: e.target.value,
+                }))
+              }
+            />
 
-              <TextField
-                color="secondary"
-                sx={{ width: '100%', marginTop: 2, marginBottom: 2 }}
-                error={isUsernameInvalid}
-                placeholder={isUsernameInvalid ? 'invalid username' : undefined}
-                label="username"
-                value={details.username}
-                onChange={(e) =>
-                  sDetails((details) => ({
-                    ...details,
-                    username: e.target.value,
-                  }))
-                }
-              />
+            <TextField
+              color="secondary"
+              sx={{ width: '100%', marginTop: 2, marginBottom: 2 }}
+              error={isUsernameInvalid}
+              placeholder={isUsernameInvalid ? 'invalid username' : undefined}
+              label="username"
+              value={details.username}
+              onChange={(e) =>
+                sDetails((details) => ({
+                  ...details,
+                  username: e.target.value,
+                }))
+              }
+            />
 
-              <SelectChipList
-                hook={foodAllergies}
-                label="allergies"
-                propertyName="allergy"
-              />
-              <SelectChipList
-                hook={medicalCond}
-                label="medical conditions"
-                propertyName="medical condition"
-              />
+            <SelectChipList
+              hook={foodAllergies}
+              label="allergies"
+              propertyName="allergy"
+            />
+            <SelectChipList
+              hook={medicalCond}
+              label="medical conditions"
+              propertyName="medical condition"
+            />
 
-              <Button
-                sx={{
-                  width: '100%',
-                  paddingTop: 2,
-                  paddingBottom: 2,
-                }}
-                color="primary"
-                variant="contained"
-                disabled={isNameInvalid || isUsernameInvalid}
-                onClick={async () => {
-                  await maybeFinishRegister();
-                }}
-              >
-                Continue
-              </Button>
-            </div>
-          </Grid>
-        </Paper>
-      </div>
-    </ThemeProvider>
+            <Button
+              sx={{
+                width: '100%',
+                paddingTop: 2,
+                paddingBottom: 2,
+              }}
+              color="primary"
+              variant="contained"
+              disabled={isNameInvalid || isUsernameInvalid}
+              onClick={async () => {
+                await maybeFinishRegister();
+              }}
+            >
+              Continue
+            </Button>
+          </div>
+        </Grid>
+      </Paper>
+    </div>
   );
 }
 
+const $wrapper: React.CSSProperties = {
+  width: '100%',
+  height: '100%',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center'
+};
 const $container: React.CSSProperties = {
   overflowY: 'auto',
   alignItems: 'center',
   justifyContent: 'center',
   width: '100%',
+  maxHeight: '90%',
   paddingRight: 10,
   alignContent: 'center',
   flexGrow: 1,
